Fix support for Azure SQL Database.
This fixes support for installing the tSQLt framework on Azure SQL Database when the collation is not the default SQL_Latin1_General_CP1_CI_AS. The problem is that the sys.* view are always SQL_Latin1_General_CP1_CI_AS on an Azure SQL database, even when the database is a different collation. This results in errors when comparing or concaternating strings where one side came from a sys.* view and the other side came from a table/view in the database. By specifying COLLATE database_default, this keeps the behaviour of on-premise SQL Server, where the sys.* views are the same collation as the database.